fixed precompile bug
